requirements.txt
requests
python-dotenv
beautifulsoup4
google-generativeai
google-cloud-bigquery
google-cloud-secret-manager
playwright
sqlalchemy
psycopg2-binary
